requirements.txt
# WebPortal Core Requirements - Production Ready
# ============================================

# Core Flask framework
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
Flask-WTF==1.2.1
WTForms==3.1.1

# Environment configuration
python-dotenv==1.0.0

# Database support
PyMySQL==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0

# Caching
redis==5.0.1

# Security and rate limiting
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
Flask-Login==0.6.3
Flask-Session==0.8.0
Flask-Admin==1.6.1

# Image processing (camera snapshots)
Pillow==10.1.0

# WebSocket client (Home Assistant)
websocket-client==1.6.4

# Production WSGI server
gunicorn==21.2.0

# System monitoring
psutil==5.9.6

# Password hashing
bcrypt==4.1.2

# Date/time utilities
python-dateutil==2.8.2

# Configuration validation
jsonschema==4.20.0

# Async HTTP client
aiohttp==3.9.1

# CLI colors and progress
colorama==0.4.6
tqdm==4.66.1

# Timezone support
pytz==2023.3.post1
